\name{SeqDef}
\alias{PlotSeqDef}
\title{Calculate and Visualize seqdef statistic}
\usage{
SeqDef(tree, df, data.col, invert, scale)

PlotSeqDef(tree, df, data.col)

}
\description{
We should add some stuff here
}

\arguments{
  \item{tree}{a phylo object}
  \item{df}{}
  \item{data.col}{a numeric value indicating the column of df to use in calculations}
  \item{invert}{}
  \item{scale}{}
}


\examples{
hello()
}



